ERR: int -> unsigned long.

+++ Code/IO/itkPNGImageIO.cxx
@@ -183,7 +183,7 @@ void PNGImageIO::Load(void* buffer)
   // update the info now that we have defined the filters
   png_read_update_info(png_ptr, info_ptr);
 
-  int rowbytes = png_get_rowbytes(png_ptr, info_ptr);
+  unsigned long rowbytes = png_get_rowbytes(png_ptr, info_ptr);
   unsigned char *tempImage = static_cast<unsigned char*>(buffer);
   png_bytep *row_pointers = new png_bytep [height];
   for (int ui = 0; ui < height; ++ui)
